introduce constants as a cc_library
