+++ constants.h
@@ -16,8 +16,7 @@
 #define CONSTANTS_H_
 
 #include <cstdint>
-
-#include "glog/logging.h"
+#include <string>
 
 namespace fido2_tests {
 // This is the status byte returned by CTAP interactions.
@@ -67,98 +66,8 @@ enum class Status : uint8_t {
   kErrOther = 0x7F
 };
 
-inline std::string StatusToString(Status status) {
-  switch (status) {
-    case Status::kErrNone:
-      return "CTAP2_OK";
-    case Status::kErrInvalidCommand:
-      return "CTAP1_ERR_INVALID_COMMAND";
-    case Status::kErrInvalidParameter:
-      return "CTAP1_ERR_INVALID_PARAMETER";
-    case Status::kErrInvalidLength:
-      return "CTAP1_ERR_INVALID_LENGTH";
-    case Status::kErrInvalidSeq:
-      return "CTAP1_ERR_INVALID_SEQ";
-    case Status::kErrTimeout:
-      return "CTAP1_ERR_TIMEOUT";
-    case Status::kErrChannelBusy:
-      return "CTAP1_ERR_CHANNEL_BUSY";
-    case Status::kErrLockRequired:
-      return "CTAP1_ERR_LOCK_REQUIRED";
-    case Status::kErrInvalidChannel:
-      return "CTAP1_ERR_INVALID_CHANNEL";
-    case Status::kErrCborUnexpectedType:
-      return "CTAP2_ERR_CBOR_UNEXPECTED_TYPE";
-    case Status::kErrInvalidCbor:
-      return "CTAP2_ERR_INVALID_CBOR";
-    case Status::kErrMissingParameter:
-      return "CTAP2_ERR_MISSING_PARAMETER";
-    case Status::kErrLimitExceeded:
-      return "CTAP2_ERR_LIMIT_EXCEEDED";
-    case Status::kErrUnsupportedExtension:
-      return "CTAP2_ERR_UNSUPPORTED_EXTENSION";
-    case Status::kErrCredentialExcluded:
-      return "CTAP2_ERR_CREDENTIAL_EXCLUDED";
-    case Status::kErrProcessing:
-      return "CTAP2_ERR_PROCESSING";
-    case Status::kErrInvalidCredential:
-      return "CTAP2_ERR_INVALID_CREDENTIAL";
-    case Status::kErrUserActionPending:
-      return "CTAP2_ERR_USER_ACTION_PENDING";
-    case Status::kErrOperationPending:
-      return "CTAP2_ERR_OPERATION_PENDING";
-    case Status::kErrNoOperations:
-      return "CTAP2_ERR_NO_OPERATIONS";
-    case Status::kErrUnsupportedAlgorithm:
-      return "CTAP2_ERR_UNSUPPORTED_ALGORITHM";
-    case Status::kErrOperationDenied:
-      return "CTAP2_ERR_OPERATION_DENIED";
-    case Status::kErrKeyStoreFull:
-      return "CTAP2_ERR_KEY_STORE_FULL";
-    case Status::kErrNoOperationPending:
-      return "CTAP2_ERR_NO_OPERATION_PENDING";
-    case Status::kErrUnsupportedOption:
-      return "CTAP2_ERR_UNSUPPORTED_OPTION";
-    case Status::kErrInvalidOption:
-      return "CTAP2_ERR_INVALID_OPTION";
-    case Status::kErrKeepaliveCancel:
-      return "CTAP2_ERR_KEEPALIVE_CANCEL";
-    case Status::kErrNoCredentials:
-      return "CTAP2_ERR_NO_CREDENTIALS";
-    case Status::kErrUserActionTimeout:
-      return "CTAP2_ERR_USER_ACTION_TIMEOUT";
-    case Status::kErrNotAllowed:
-      return "CTAP2_ERR_NOT_ALLOWED";
-    case Status::kErrPinInvalid:
-      return "CTAP2_ERR_PIN_INVALID";
-    case Status::kErrPinBlocked:
-      return "CTAP2_ERR_PIN_BLOCKED";
-    case Status::kErrPinAuthInvalid:
-      return "CTAP2_ERR_PIN_AUTH_INVALID";
-    case Status::kErrPinAuthBlocked:
-      return "CTAP2_ERR_PIN_AUTH_BLOCKED";
-    case Status::kErrPinNotSet:
-      return "CTAP2_ERR_PIN_NOT_SET";
-    case Status::kErrPinRequired:
-      return "CTAP2_ERR_PIN_REQUIRED";
-    case Status::kErrPinPolicyViolation:
-      return "CTAP2_ERR_PIN_POLICY_VIOLATION";
-    case Status::kErrPinTokenExpired:
-      return "CTAP2_ERR_PIN_TOKEN_EXPIRED";
-    case Status::kErrRequestTooLarge:
-      return "CTAP2_ERR_REQUEST_TOO_LARGE";
-    case Status::kErrActionTimeout:
-      return "CTAP2_ERR_ACTION_TIMEOUT";
-    case Status::kErrUpRequired:
-      return "CTAP2_ERR_UP_REQUIRED";
-    case Status::kErrUvBlocked:
-      return "CTAP2_ERR_UV_BLOCKED";
-    case Status::kErrOther:
-      return "CTAP1_ERR_OTHER";
-    default:
-      CHECK(false) << "unreachable default - TEST SUITE BUG";
-  }
-}
+// Converts a Status to a string for printing.
+std::string StatusToString(Status status);
 
 // These are the possible CTAP commands.
 enum class Command : uint8_t {
@@ -169,6 +78,9 @@ enum class Command : uint8_t {
   kAuthenticatorReset = 0x07,
   kAuthenticatorGetNextAssertion = 0x08
 };
+
+// Converts a Command to a string for printing.
+std::string CommandToString(Command command);
 
 // ES256 and RS256 are for signatures, while ECDH is for key agreement.
 enum class Algorithm {
